requirements.txt
pylint
flake8
gitpython
google-genai
python-dotenv
